requirements.txt
# Financial Advisor Multi-Agent System Requirements
# Core dependencies for the AutoGen-based stock analysis system

# Data sources
alpha_vantage>=2.3.1

# AutoGen framework with OpenAI and Azure support
autogen-ext[openai,azure]>=0.4.0

# Environment management
python-dotenv>=1.0.0

# Data processing and visualization
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0

# OpenAI client (for direct OpenAI usage)
openai>=1.0.0

# Web service framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Additional utilities for testing
requests>=2.31.0

# Additional utilities (automatically installed with above packages)
# requests - HTTP requests
# asyncio - Async support (built-in Python 3.7+)
# typing - Type hints (built-in Python 3.5+)
